LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-inert-invoker-expected.txt

Before Invoker Inside Popover After Popover

FAIL Popover focus navigation with an inert invoker assert_equals: Focus should move into the popover expected Element node <button id="inside">Inside Popover</button> but got Element node <button id="before">Before</button>
